added inputs from request

+++ src/main/java/com/aiplatform/tracking_service/dto/DietSummary.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class DietSummary {
-    private int todayCalories;
+    private double todayCalories;
     private int goal;
-    private int entriesCount;
+    private long entriesCount;
 }
